pyproject.toml
[tool.poetry]
name = "ai-trader-mt5"
version = "0.1.0"
description = "AI Trading Robot for MetaTrader 5"
authors = ["DiamondDim <your-email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
MetaTrader5 = "5.0.5328"
pandas = "^2.0.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
PyYAML = "^6.0"
joblib = "^1.3.0"
scipy = "^1.11.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
